clio-kit-mcp-servers/parallel-sort/pyproject.toml
[project]
name = "parallel-sort-mcp"
version = "1.0.0"
description = "Parallel Sort MCP - High-Performance Log File Processing for LLMs with advanced sorting and analysis"
readme = "README.md"
requires-python = ">=3.10"
license = "BSD-3-Clause"
authors = [
    { name = "IoWarp Team - Gnosis Research Center", email = "grc@illinoistech.edu" }
]

keywords = ["parallel-sorting", "log-processing", "log-analysis", "high-performance", "timestamp-sorting", "pattern-detection", "log-filtering", "data-export"]

dependencies = [
  "fastmcp>=2.13.0",
  "python-dotenv>=1.0.0",
  "pandas>=1.5.0",
  "aiofiles>=23.0.0"
]

[project.urls]
Homepage = "https://toolkit.iowarp.ai/"
Repository = "https://github.com/iowarp/clio-kit"

[project.scripts]
parallel-sort-mcp = "server:main"

[dependency-groups]
dev = [
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.12.5",
    "types-aiofiles>=23.0.0",
]

[build-system]
requires = ["setuptools>=64.0", "wheel"]
build-backend = "setuptools.build_meta"
